Move document request types to common
